move BackendConnection to jdbc.connection package
