+++ sharding-proxy/src/main/java/io/shardingsphere/proxy/backend/BackendHandlerFactory.java
@@ -18,7 +18,7 @@
 package io.shardingsphere.proxy.backend;
 
 import io.shardingsphere.core.constant.DatabaseType;
-import io.shardingsphere.proxy.backend.jdbc.BackendConnection;
+import io.shardingsphere.proxy.backend.jdbc.connection.BackendConnection;
 import io.shardingsphere.proxy.backend.jdbc.JDBCBackendHandler;
 import io.shardingsphere.proxy.backend.jdbc.execute.JDBCExecuteEngineFactory;
 import io.shardingsphere.proxy.backend.netty.SQLPacketsBackendHandler;
