+++ sharding-proxy/src/main/java/io/shardingsphere/proxy/backend/jdbc/execute/JDBCExecuteEngineFactory.java
@@ -18,7 +18,7 @@
 package io.shardingsphere.proxy.backend.jdbc.execute;
 
 import io.shardingsphere.proxy.backend.constant.ProxyMode;
-import io.shardingsphere.proxy.backend.jdbc.BackendConnection;
+import io.shardingsphere.proxy.backend.jdbc.connection.BackendConnection;
 import io.shardingsphere.proxy.backend.jdbc.execute.memory.ConnectionStrictlyExecuteEngine;
 import io.shardingsphere.proxy.backend.jdbc.execute.stream.MemoryStrictlyExecuteEngine;
 import io.shardingsphere.proxy.backend.jdbc.wrapper.JDBCExecutorWrapper;
